+++ Nesting.cpp
@@ -5,7 +5,7 @@ int Nesting(vector<int> &S) {
     return 0;
   }
   stack<char> stack;
-  for (int i = 0; i < S.size(); i++) {
+  for (size_t i = 0; i < S.size(); i++) {
     if (S[i] == '(') {
       stack.push(S[i]);
     }
