+++ OddOccurrencesinArray.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int OddOcurrencesinArray(vector<int> &A) {
   int result = 0;
-  for (int i = 0; i < A.size(); i++) {
+  for (size_t i = 0; i < A.size(); i++) {
     result ^= A[i];
   }
   return result;
